# Applying Coroutines to Show/Hide CDT and Aura Frames