--- a/RaidFrameSettings.toc
+++ b/RaidFrameSettings.toc
@@ -28,6 +28,7 @@ Utils\ProfileSwitcher.lua
 Utils\CooldownText.lua
 Utils\Helper.lua
 Utils\HookRegistry.lua
+Utils\Queue.lua
 
 #Modules
 modules.xml