This is not real PR.
I just want to share what I've been testing.

I tested applying a coroutine to the cooldown timer and show/hide of the aura frame to reduce lag.

The results were that the Cooldown Timer had a big effect, but the show/hide in aura, it actually decreased the FPS even more.
However, I'm hopeful that the lag was reduced at the sacrifice of more frames.

When applying the coroutine to the Cooldown Timer, I think it would be more effective to have it pause for 0.1~0.5 seconds after looping once and then loop again. 


1. https://youtu.be/bZSY3udZ1Sc - 19.6ms | Apply coroutines to CDTs
2. https://youtu.be/OF8lcL6N7WY - 24.3ms | Apply coroutines to CDTs and aura queue
3. https://youtu.be/5F1bgKxm0bE - 55.54ms | Original RFS
4. https://youtu.be/ss8-lFE6e6s - 57.3ms | Apply coroutines to aura queue



Oh, by the way, to create an extreme environment, I tested with 30 buffs and 15 debuffs in each frame. This would never happen in the real world.