@Slothpala 2024-04-12
Throttling may be a better way to go than coroutines.