@excorp 2024-04-12
I've said it before, but I'm a novice at addon development. 
I don't even know lua very well. I found out about coroutines by some chance, 
and I thought that using them to time-slice would prevent one routine from taking up a lot of CPU time, 
and that would reduce the freezing (lag), so I tried it. 

But I think the cooldown timer won't take up a lot of CPU time, 
so just giving it an interval like you said should be enough.

I just wanted to share that I tested this!  😃