@Slothpala 2024-04-12
Yeah, it's fun to play around with :D I've read that corotinues in wow lua use a lot of memory, but I haven't tested it.


I also measured some extremely high CPU usage (with the AddOns CPU Usage and Addon Usage add-ons) when in raid (Amirdrassil), but the FPS drop was only 3 to 5 fps and no noticeable lag on either my low spec Macbook or my gaming PC. 

So I made a small test addon that just hooks into the functions that RFS uses and does nothing but call a function that adds two numbers. And I also got this extreme CPU usage.
So I am not sure if these addons (or rather the functions they call to get their data) are really suitable for measuring performance. My suspicion is that some of the work done by the built-in Blizzard functions is being blamed on the addon. The strange thing though is that this only happens when in the raid. In epic bgs with 40 man in my group and all modules enabled the usage is around 30 to 40ms which sounds about right for what the addon does.