@Slothpala 2024-04-12
local function add_two_numbers(nbr1, nbr2)
    local result = nbr1 + nbr2
    return result
end

hooksecurefunc("CompactUnitFrame_UpdateAuras", function ()
    local result = add_two_numbers(8, 9)
end)

hooksecurefunc("CompactUnitFrame_HideAllBuffs", function ()
    local result = add_two_numbers(4, 9)
end)

hooksecurefunc("CompactUnitFrame_HideAllDebuffs", function ()
    local result = add_two_numbers(2, 3)
end)

hooksecurefunc("DefaultCompactUnitFrameSetup", function ()
    local result = add_two_numbers(4, 7)
end)

hooksecurefunc("CompactUnitFrame_UtilSetBuff", function ()
    local result = add_two_numbers(1, 4)
end)

hooksecurefunc("CompactUnitFrame_UtilSetDebuff", function ()
    local result = add_two_numbers(9, 2)
end)