@Slothpala 2024-04-12
With this addon I get 200 to 400 ms of addon usage, which sounds a bit high for adding some numbers. So I am a bit confused, is securehooking just that expensive or is the work of Blizzard's own functions measured when hooking.