@excorp 2024-04-14
It's amazing how simply adding two numbers can increase by 200-400ms.
I'm closing this PR anyway